Crate uds

A unix domain sockets library that supports abstract addresses, fd-passing, SOCK_SEQPACKET sockets and more.

File-descriptor passing and abstract socket support for stream and datagram sockets is provided via extension traits for existing types in std::os::unix::net and from mio_uds (the latter is opt-in and must be enabled with features=["mio_uds"] in Cargo.toml).

See README for status of operating system support and other general info.

Modules

nonblocking

Structs

UnixSeqpacketConn

An unix domain sequential packet connection.

UnixSeqpacketListener

An unix domain listener for sequential packet connections.

UnixSocketAddr

A unix domain socket address.

Enums

AddrName

An enum representation of an unix socket address.

ConnCredentials

Credentials of the peer process when it called connect(), accept() or pair().

Traits

UnixDatagramExt

Extension trait for std::os::unix::net::UnixDatagram and nonblocking equivalents.

UnixListenerExt

Extension trait for using UnixSocketAddr with UnixListener types.

UnixStreamExt

Extension trait for std::os::unix::net::UnixDatagram and nonblocking equivalents.

Type Definitions

UnixSocketAddrRef