Struct uds::UnixSeqpacketConn

#[repr(transparent)]
pub struct UnixSeqpacketConn { /* private fields */ }

An unix domain sequential packet connection.

Sequential-packet connections have an interface similar to streams, but behave more like connected datagram sockets.

They have guaranteed in-order and reliable delivery, which unix datagrams technically doesn’t.

Operating system support

Sequential-packet sockets are supported by Linux, FreeBSD, NetBSD and Illumos, but not by for example macOS or OpenBSD.

Zero-length packets

… are best avoided:
On Linux and FreeBSD zero-length packets can be sent and received, but there is no way to distinguish receiving one from reaching end of connection unless the packet has an ancillary payload. Also beware of trying to receive with a zero-length buffer, as that will on FreeBSD (and probably other BSDs with seqpacket sockets) always succeed even if there is no packet waiting.

Illumos and Solaris doesn’t support receiving zero-length packets at all: writes succeed but recv() will block.

Examples

What is sent separately is received separately:

let (a, b) = uds::UnixSeqpacketConn::pair().expect("Cannot create seqpacket pair");

a.send(b"first").unwrap();
a.send(b"second").unwrap();

let mut buffer_big_enough_for_both = [0; 20];
let (len, _truncated) = b.recv(&mut buffer_big_enough_for_both).unwrap();
assert_eq!(&buffer_big_enough_for_both[..len], b"first");
let (len, _truncated) = b.recv(&mut buffer_big_enough_for_both).unwrap();
assert_eq!(&buffer_big_enough_for_both[..len], b"second");

Connect to a listener on a socket file and write to it:

use uds::{UnixSeqpacketListener, UnixSeqpacketConn};

let listener = UnixSeqpacketListener::bind("seqpacket.socket")
    .expect("create seqpacket listener");
let conn = UnixSeqpacketConn::connect("seqpacket.socket")
    .expect("connect to seqpacket listener");

let message = "Hello, listener";
let sent = conn.send(message.as_bytes()).unwrap();
assert_eq!(sent, message.len());

std::fs::remove_file("seqpacket.socket").unwrap(); // clean up after ourselves

Connect to a listener on an abstract address:

use uds::{UnixSeqpacketListener, UnixSeqpacketConn, UnixSocketAddr};

let addr = UnixSocketAddr::new("@seqpacket example").unwrap();
let listener = UnixSeqpacketListener::bind_unix_addr(&addr)
    .expect("create abstract seqpacket listener");
let _client = UnixSeqpacketConn::connect_unix_addr(&addr)
    .expect("connect to abstract seqpacket listener");
let (_server, _addr) = listener.accept_unix_addr().unwrap();

Implementations

impl UnixSeqpacketConn

pub fn connect<P: AsRef<Path>>(path: P) -> Result<Self, Error>

Connects to an unix seqpacket server listening at path.

This is a wrapper around connect_unix_addr() for convenience and compatibility with std.

pub fn connect_unix_addr(addr: &UnixSocketAddr) -> Result<Self, Error>

Connects to an unix seqpacket server listening at addr.

pub fn connect_from_to_unix_addr(
from: &UnixSocketAddr,
to: &UnixSocketAddr
) -> Result<Self, Error>

Binds to an address before connecting to a listening seqpacet socket.

pub fn pair() -> Result<(Self, Self), Error>

Creates a pair of unix-domain seqpacket conneections connected to each other.

Examples

let (a, b) = uds::UnixSeqpacketConn::pair().unwrap();
assert!(a.local_unix_addr().unwrap().is_unnamed());
assert!(b.local_unix_addr().unwrap().is_unnamed());

a.send(b"hello").unwrap();
b.recv(&mut[0; 20]).unwrap();

pub fn local_unix_addr(&self) -> Result<UnixSocketAddr, Error>

Returns the address of this side of the connection.

pub fn peer_unix_addr(&self) -> Result<UnixSocketAddr, Error>

Returns the address of the other side of the connection.

pub fn initial_peer_credentials(&self) -> Result<ConnCredentials, Error>

Returns information about the process of the peer when the connection was established.

See documentation of the returned type for details.

pub fn initial_peer_selinux_context(
&self,
buf: &mut [u8]
) -> Result<usize, Error>

Returns the SELinux security context of the process that created the other end of this connection.

Will return an error on other operating systems than Linux or Android, and also if running inside kubernetes. On success the number of bytes used is returned. (like Read)

The default security context is unconfined, without any trailing NUL.
A buffor of 50 bytes is probably always big enough.

pub fn send(&self, packet: &[u8]) -> Result<usize, Error>

Sends a packet to the peer.

pub fn recv(&self, buffer: &mut [u8]) -> Result<(usize, bool), Error>

Receives a packet from the peer.

The returned bool indicates whether the packet was truncated due to the buffer beeing too small.

pub fn send_vectored(&self, slices: &[IoSlice<'_>]) -> Result<usize, Error>

Sends a packet assembled from multiple byte slices.

pub fn recv_vectored(
&self,
buffers: &mut [IoSliceMut<'_>]
) -> Result<(usize, bool), Error>

Reads a packet into multiple buffers.

The returned bool indicates whether the packet was truncated due to too short buffers.

pub fn send_fds(&self, bytes: &[u8], fds: &[RawFd]) -> Result<usize, Error>

Sends a packet with associated file descriptors.

pub fn recv_fds(
&self,
byte_buffer: &mut [u8],
fd_buffer: &mut [RawFd]
) -> Result<(usize, bool, usize), Error>

Receives a packet and associated file descriptors.

pub fn peek(&self, buffer: &mut [u8]) -> Result<(usize, bool), Error>

Receives a packet without removing it from the incoming queue.

The returned bool indicates whether the packet was truncated due to the buffer being too small.

Examples

let (a, b) = uds::UnixSeqpacketConn::pair().unwrap();
a.send(b"hello").unwrap();
let mut buf = [0u8; 10];
assert_eq!(b.peek(&mut buf[..1]).unwrap(), (1, true));
assert_eq!(&buf[..2], &[b'h', 0]);
assert_eq!(b.peek(&mut buf).unwrap(), (5, false));
assert_eq!(&buf[..5], b"hello");
assert_eq!(b.recv(&mut buf).unwrap(), (5, false));
assert_eq!(&buf[..5], b"hello");

pub fn peek_vectored(
&self,
buffers: &mut [IoSliceMut<'_>]
) -> Result<(usize, bool), Error>

Receives a packet without removing it from the incoming queue.

The returned bool indicates whether the packet was truncated due to the combined buffers being too small.

pub fn take_error(&self) -> Result<Option<Error>, Error>

Returns the value of the SO_ERROR option.

This might only provide errors generated from nonblocking connect()s, which this library doesn’t support. It is therefore unlikely to be useful, but is provided for parity with stream counterpart in std.

Examples

let (a, b) = uds::UnixSeqpacketConn::pair().unwrap();
drop(b);

assert!(a.send(b"anyone there?").is_err());
assert!(a.take_error().unwrap().is_none());

pub fn try_clone(&self) -> Result<Self, Error>

Creates a new file descriptor also pointing to this side of this connection.

Examples

Both new and old can send and receive, and share queues:

let (a1, b) = uds::nonblocking::UnixSeqpacketConn::pair().unwrap();
let a2 = a1.try_clone().unwrap();

a1.send(b"first").unwrap();
a2.send(b"second").unwrap();

let mut buf = [0u8; 20];
let (len, _truncated) = b.recv(&mut buf).unwrap();
assert_eq!(&buf[..len], b"first");
b.send(b"hello first").unwrap();
let (len, _truncated) = b.recv(&mut buf).unwrap();
assert_eq!(&buf[..len], b"second");
b.send(b"hello second").unwrap();

let (len, _truncated) = a2.recv(&mut buf).unwrap();
assert_eq!(&buf[..len], b"hello first");
let (len, _truncated) = a1.recv(&mut buf).unwrap();
assert_eq!(&buf[..len], b"hello second");

Clone can still be used after the first one has been closed:

let (a, b1) = uds::nonblocking::UnixSeqpacketConn::pair().unwrap();
a.send(b"hello").unwrap();

let b2 = b1.try_clone().unwrap();
drop(b1);
assert_eq!(b2.recv(&mut[0; 10]).unwrap().0, "hello".len());

pub fn set_read_timeout(&self, timeout: Option<Duration>) -> Result<(), Error>

Sets the read timeout to the duration specified.

If the value specified is None, then recv() and its variants will block indefinitely.
An error is returned if the duration is zero.

The duration is rounded to microsecond precission. Currently it’s rounded down except if that would make it all zero.

Operating System Support

On Illumos (and pressumably also Solaris) timeouts appears not to work for unix domain sockets.

Examples

use std::io::ErrorKind;
use std::time::Duration;
use uds::UnixSeqpacketConn;

let (a, b) = UnixSeqpacketConn::pair().unwrap();
a.set_read_timeout(Some(Duration::new(0, 2_000_000))).unwrap();
let error = a.recv(&mut[0; 1024]).unwrap_err();
assert_eq!(error.kind(), ErrorKind::WouldBlock);

pub fn read_timeout(&self) -> Result<Option<Duration>, Error>

Returns the read timeout of this socket.

None is returned if there is no timeout.

Note that subsecond parts might have been be rounded by the OS (in addition to the rounding to microsecond in set_read_timeout()).

Examples

use uds::UnixSeqpacketConn;
use std::time::Duration;

let timeout = Some(Duration::new(2, 0));
let conn = UnixSeqpacketConn::pair().unwrap().0;
conn.set_read_timeout(timeout).unwrap();
assert_eq!(conn.read_timeout().unwrap(), timeout);

pub fn set_write_timeout(&self, timeout: Option<Duration>) -> Result<(), Error>

Sets the write timeout to the duration specified.

If the value specified is None, then send() and its variants will block indefinitely.
An error is returned if the duration is zero.

Operating System Support

On Illumos (and pressumably also Solaris) timeouts appears not to work for unix domain sockets.

Examples

let (conn, _other) = UnixSeqpacketConn::pair().unwrap();
conn.set_write_timeout(Some(Duration::new(0, 500 * 1000))).unwrap();
loop {
    if let Err(e) = conn.send(&[0; 1000]) {
        assert_eq!(e.kind(), ErrorKind::WouldBlock, "{}", e);
        break
    }
}

pub fn write_timeout(&self) -> Result<Option<Duration>, Error>

Returns the write timeout of this socket.

None is returned if there is no timeout.

Examples

let conn = uds::UnixSeqpacketConn::pair().unwrap().0;
assert!(conn.write_timeout().unwrap().is_none());

pub fn set_nonblocking(&self, nonblocking: bool) -> Result<(), Error>

Enables or disables nonblocking mode.

Consider using the nonblocking variant of this type instead. This method mainly exists for feature parity with std’s UnixStream.

Examples

Trying to receive when there are no packets waiting:

let (a, b) = UnixSeqpacketConn::pair().expect("create seqpacket pair");
a.set_nonblocking(true).unwrap();
assert_eq!(a.recv(&mut[0; 20]).unwrap_err().kind(), ErrorKind::WouldBlock);

Trying to send when the OS buffer for the connection is full:

let (a, b) = UnixSeqpacketConn::pair().expect("create seqpacket pair");
a.set_nonblocking(true).unwrap();
loop {
    if let Err(error) = a.send(&[b'#'; 1000]) {
        assert_eq!(error.kind(), ErrorKind::WouldBlock);
        break;
    }
}

pub fn shutdown(&self, how: Shutdown) -> Result<()>

Shuts down the read, write, or both halves of this connection.