Struct uds::UnixSeqpacketListener

#[repr(transparent)]
pub struct UnixSeqpacketListener { /* private fields */ }

An unix domain listener for sequential packet connections.

See UnixSeqpacketConn for a description of this type of connection.

Examples

let listener = uds::UnixSeqpacketListener::bind("seqpacket_listener.socket")
    .expect("Create seqpacket listener");
let _client = uds::UnixSeqpacketConn::connect("seqpacket_listener.socket").unwrap();
let (conn, _addr) = listener.accept_unix_addr().unwrap();
conn.send(b"Welcome").unwrap();

Implementations

impl UnixSeqpacketListener

pub fn bind<P: AsRef<Path>>(path: P) -> Result<Self, Error>

Creates a socket that listens for seqpacket connections on the specified socket file.

pub fn bind_unix_addr(addr: &UnixSocketAddr) -> Result<Self, Error>

Creates a socket that listens for seqpacket connections on the specified address.

pub fn local_unix_addr(&self) -> Result<UnixSocketAddr, Error>

Returns the address the socket is listening on.

pub fn accept_unix_addr(
&self
) -> Result<(UnixSeqpacketConn, UnixSocketAddr), Error>

Accepts a new incoming connection to this listener.

pub fn take_error(&self) -> Result<Option<Error>, Error>

Returns the value of the SO_ERROR option.

This might never produce any errors for listeners. It is therefore unlikely to be useful, but is provided for parity with std::unix::net::UnixListener.

pub fn try_clone(&self) -> Result<Self, Error>

Creates a new file descriptor listening for the same connections.

pub fn set_timeout(&self, timeout: Option<Duration>) -> Result<(), Error>

Sets a maximum duration to wait in a single accept() on this socket.

None disables a previously set timeout. An error is returned if the duration is zero.

Operating System Support

Only Linux appers to apply timeouts to accept().
On macOS, FreeBSD and NetBSD, timeouts are silently ignored.
On Illumos setting timeouts for all unix domain sockets silently fails.

On OSes where timeouts are known to not work, this function will return an error even if setting the timeout didn’t fail.

Examples

let listener = UnixSeqpacketListener::bind_unix_addr(&addr).unwrap();
listener.set_timeout(Some(Duration::new(0, 200_000_000))).unwrap();
let err = listener.accept_unix_addr().unwrap_err();
assert_eq!(err.kind(), ErrorKind::WouldBlock);

pub fn timeout(&self) -> Result<Option<Duration>, Error>

Returns the timeout for accept() on this socket.

None is returned if there is no timeout.

Even if a timeout has is set, it is ignored by accept() on most operating systems except Linux.

Examples

let listener = UnixSeqpacketListener::bind_unix_addr(&addr).unwrap();
assert_eq!(listener.timeout().unwrap(), None);
let timeout = Some(Duration::new(2, 0));
listener.set_timeout(timeout).unwrap();
assert_eq!(listener.timeout().unwrap(), timeout);

pub fn set_nonblocking(&self, nonblocking: bool) -> Result<(), Error>

Enables or disables nonblocking-ness of [accept_unix_addr()](#method.accept_unix addr).

The returned connnections will still be in blocking mode regardsless.

Consider using the nonblocking variant of this type instead; this method mostly exists for feature parity with std’s UnixListener.

Examples

let listener = UnixSeqpacketListener::bind_unix_addr(&addr).expect("create listener");
listener.set_nonblocking(true).expect("enable noblocking mode");
assert_eq!(listener.accept_unix_addr().unwrap_err().kind(), ErrorKind::WouldBlock);