Trait uds::UnixStreamExt

pub trait UnixStreamExt: AsRawFd + FromRawFd + Sized {
    fn connect_to_unix_addr(addr: &UnixSocketAddr) -> Result<Self, Error>;
    fn connect_from_to_unix_addr(
        from: &UnixSocketAddr,
        to: &UnixSocketAddr
    ) -> Result<Self, Error>;

    fn local_unix_addr(&self) -> Result<UnixSocketAddr, Error> { ... }
    fn peer_unix_addr(&self) -> Result<UnixSocketAddr, Error> { ... }
    fn send_fds(&self, bytes: &[u8], fds: &[RawFd]) -> Result<usize, Error> { ... }
    fn recv_fds(
        &self,
        buf: &mut [u8],
        fd_buf: &mut [RawFd]
    ) -> Result<(usize, usize), Error> { ... }
    fn initial_peer_credentials(&self) -> Result<ConnCredentials, Error> { ... }
    fn initial_peer_selinux_context(
        &self,
        buffer: &mut [u8]
    ) -> Result<usize, Error> { ... }
}

Extension trait for std::os::unix::net::UnixDatagram and nonblocking equivalents.

Required Methods

fn connect_to_unix_addr(addr: &UnixSocketAddr) -> Result<Self, Error>

Creates a connection to a listening path-based or abstract named socket.

fn connect_from_to_unix_addr(
from: &UnixSocketAddr,
to: &UnixSocketAddr
) -> Result<Self, Error>

Creates a path-based or abstract-named socket and connects to a listening socket.

Provided Methods

fn local_unix_addr(&self) -> Result<UnixSocketAddr, Error>

Get the address of this socket, as a type that fully supports abstract addresses.

fn peer_unix_addr(&self) -> Result<UnixSocketAddr, Error>

Returns the address of the other end of this stream, as a type that fully supports abstract addresses.

fn send_fds(&self, bytes: &[u8], fds: &[RawFd]) -> Result<usize, Error>

Sends file descriptors in addition to bytes.

fn recv_fds(
&self,
buf: &mut [u8],
fd_buf: &mut [RawFd]
) -> Result<(usize, usize), Error>

Receives file descriptors in addition to bytes.

fn initial_peer_credentials(&self) -> Result<ConnCredentials, Error>

Returns the credentials of the process that created the other end of this stream.

fn initial_peer_selinux_context(&self, buffer: &mut [u8]) -> Result<usize, Error>

Returns the SELinux security context of the process that created the other end of this stream.

Will return an error on other operating systems than Linux or Android, and also if running under kubernetes. On success the number of bytes used is returned. (like Read)

The default security context is unconfined, without any trailing NUL.
A buffor of 50 bytes is probably always big enough.